# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fursa Platform API - Digital Employment Platform for Gaza
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List approved & open jobs (public)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
            enum: [online, field, hybrid]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJobListResponse"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Public job detail (approved & open only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /me/saved-jobs/{id}:
    post:
      operationId: saveJob
      tags: [seeker]
      summary: Save a job to favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Saved
    delete:
      operationId: unsaveJob
      tags: [seeker]
      summary: Remove saved job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /jobs/{id}/apply:
    post:
      operationId: applyToJob
      tags: [seeker]
      summary: Submit an application to a job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyApplication"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    JobType:
      type: string
      enum: [online, field, hybrid]
...
    ApplicationStatus:
      type: string
      enum: [pending, accepted, rejected]
...
    PublicJob:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/JobType"
        category:
          type: string
        employerName:
          type: string
        employerLocation:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - category
        - employerName
        - createdAt
    PublicJobDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/JobType"
        category:
          type: string
        contactInfo:
          type: string
        employerName:
          type: string
        employerLocation:
          type: ["string", "null"]
        employerBio:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
          format: date-time
        isOpen:
          type: boolean
        createdAt:
          type: string
          format: date-time
        savedByMe:
          type: boolean
        appliedByMe:
          type: boolean
      required:
        - id
        - title
        - description
        - type
        - category
        - contactInfo
        - employerName
        - isOpen
        - createdAt
        - savedByMe
        - appliedByMe
    PublicJobListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicJob"
        total:
          type: integer
      required: [items, total]
...
    CreateApplicationBody:
      type: object
      properties:
        coverLetter:
          type: ["string", "null"]
        cvObjectPath:
          type: ["string", "null"]
    MyApplication:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: string
        employerName:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        coverLetter:
          type: ["string", "null"]
        cvObjectPath:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, jobId, jobTitle, employerName, status, createdAt]
...