# lib/api-zod/src/generated/types/publicJobDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
import type { JobType } from "./jobType";

export interface PublicJobDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  type: JobType;
  category: string;
  contactInfo: string;
  employerName: string;
  /** @nullable */
  employerLocation?: string | null;
  /** @nullable */
  employerBio?: string | null;
  /** @nullable */
  deadline?: Date | null;
  isOpen: boolean;
  createdAt: Date;
  savedByMe: boolean;
  appliedByMe: boolean;
}