# lib/api-zod/src/generated/types/publicJobListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
import type { PublicJob } from "./publicJob";

export interface PublicJobListResponse {
  items: PublicJob[];
  total: number;
}